Share one compile cache across the Java suites
The spec, e2e and wasi-testsuite suites each carried their own copy of
the same content-addressed compile-and-rename recipe, and the copies
disagreed: two cache prefixes inside /tmp/dewasm-java-cache plus a
third cache directory of its own for spec, so a source two suites
agreed on was compiled once per suite. memory_grow had no cache at
all — it compiled into a fresh per-pid scratch dir on every run and
never removed it, leaving a trail of directories in /tmp.

Move the recipe into tests/common/mod.rs under one prefix in one
directory and point all four suites at it.

Measured on the Java fast gate with a cold compile cache: user CPU
247.0 s -> 245.7 s, wall 68 s -> 61 s. The dedup itself buys nothing
at the fast tier (166 trials produce 166 distinct sources); the win is
one recipe instead of four, plus caching and the /tmp leak for
memory_grow.
